ADDED: time log table(unfinished)

+++ src/main/java/cz/muni/fi/pv168/project/ui/LoginWindow.java
@@ -85,7 +85,7 @@ public class LoginWindow {
     }
 
     private void login() {
-        MainWindow mainWindow = new MainWindow(new User(usernameField.getText(), idField.getValue()));
+        MainWindow mainWindow = new MainWindow(new User(usernameField.getText(), (long) idField.getValue()));
         System.out.println("logged as: " + usernameField.getText());
         mainWindow.show();
     }
